+++ src/const.js
@@ -1,2 +1,4 @@
 export default const effect_flat_footed = "Compendium.pf2e-action-support.action-support.Item.Vh5E1Qgp34sTKfVs";
 export const effect_panache = "Compendium.pf2e.feat-effects.Item.uBJsxCzNhje8m8jj";
+export const effect_disarm_success = "Compendium.pf2e.equipment-effects.Item.z3ATL8DcRVrT0Uzt";
+export const effect_adverse_subsist_situation = "Compendium.pf2e.other-effects.Item.wHWWHkjDXmJl4Ia6";