+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Version 0.0.3
+
+### New feature
+- Add support Disarm Action
+- Add support Feint Action
+- Add support High jump Action
+- Add support Long jump Action
+- Add support Shove Action
+- Add support Subsist Action
+- Add support Tamper Action
+- Add support Grapple Action
+
 ## Version 0.0.2
 
 ### New feature